Schedule controller rewritten and added corresponded interfaces and implementations to BL

+++ SofthemeClassrooms/DataAccessLayer/DataBaseRepositories.cs
@@ -82,7 +82,7 @@ namespace DataAccessLayer
             }
         }
 
-        public IRepository<Event, int> Events
+        public IRepository<Event, long> Events
         {
             get
             {
